background behaviors refactor: (fixes #23) - move auto-play, auto-fetch and auto-scroll behaviors to behaviors/global/* - bgbehaviors manages these background behaviors - command line --bgbehaviors option specifies which background behaviors to run (defaults to auto-fetch and auto-play)
