+++ Dockerfile
@@ -35,6 +35,7 @@ RUN yarn install
 ADD config.yaml /app/
 ADD uwsgi.ini /app/
 ADD *.js /app/
+ADD behaviors/ /app/behaviors/
 
 RUN ln -s /app/main.js /usr/bin/crawl
